### 2018-04-18 #681 @whedon
[PRE REVIEW]: BeeNestABM: An open-source agent-based model of spatiotemporal distribution of bumblebees in nests
**Submitting author:** @ashleefv (<a href="http://orcid.org/0000-0001-9059-5703">Ashlee N. Ford Versypt</a>)
**Repository:** <a href="https://github.com/ashleefv/BeeNestABM" target ="_blank">https://github.com/ashleefv/BeeNestABM</a>
**Version:** 1.0
