+++ paper.bib
@@ -1,11 +1,11 @@
 @Article{Crall2018b,
-  AUTHOR = {Crall, J. A. and de Bivort, B. L. and Dey, B. and Ford Versypt, A. N.},
+  AUTHOR = {Crall, J. A. and deBivort, B. L. and Dey, B. and Ford Versypt, A. N.},
   TITLE = {{Social buffering of pesticides in bumblebees: agent-based modeling of the effects of colony size and neonicotinoid exposure on nest behavior}},
   YEAR = {2018},
   JOURNAL = {Submitted to Frontiers in Ecology and Evolution}
 }
 @Article{Crall2018a,
-  AUTHOR = {Crall, J. A. and Switzer, C. and Oppenheimer, R. and Combes, S. and Pierce, N. and De Bivort, B. and Ford Versypt, A. N. and Dey, B. and Brown, A. and Eyster, M. and Guerin, C.},
+  AUTHOR = {Crall, J. A. and Switzer, C. and Oppenheimer, R. and Combes, S. and Pierce, N. and de Bivort, B. and Ford Versypt, A. N. and Dey, B. and Brown, A. and Eyster, M. and Guerin, C.},
   TITLE = {{Chronic neonicotinoid exposure disrupts bumblebee nest behavior, social networks, and thermoregulation
 }},
   YEAR = {2018},
