### 2018-04-27 #681 @Kevin-Mattheus-Moerman
@pboesu @giskid are you able to review this submission for [JOSS](http://joss.theoj.org/about)?

[--> Software repository :floppy_disk: : <--](https://github.com/ashleefv/BeeNestABM)

[--> Article proof :page_facing_up: <--](https://github.com/openjournals/joss-papers/blob/joss.00681/joss.00681/10.21105.joss.00681.pdf)
