### 2018-04-30 #681 @GISKid
I cannot, although looks interesting!

On Fri, Apr 27, 2018, 11:38 AM Kevin Mattheus Moerman, <
notifications@github.com> wrote:

> @pboesu <https://github.com/pboesu> @GISKid <https://github.com/GISKid>
> are you able to review this submission for JOSS
> <http://joss.theoj.org/about>?
>
> --> Software repository 💾 : <-- <https://github.com/ashleefv/BeeNestABM>
>
> --> Article proof 📄 <--
> <https://github.com/openjournals/joss-papers/blob/joss.00681/joss.00681/10.21105.joss.00681.pdf>
>
