### 2018-05-03 #718 @Kevin-Mattheus-Moerman
@ashleefv for the `submitted` works cited in your paper, are you able to post these on a pre-print server and add the URL's or DOI's to these references? e.g. https://engrxiv.org/ (you can check your rights for individual journals [here](http://www.sherpa.ac.uk/romeo/index.php))
