### 2018-05-03 ba78b1a
removed self-citation line
removed reference to "submitted" works that are under review.
