### 2018-05-03 #718 @ashleefv
I removed the line referencing the submitted works. My co-authors preferred
this over pre-prints after the papers are in review.

Ashlee N. Ford Versypt, Ph.D.
Assistant Professor, School of Chemical Engineering, Oklahoma State
University
Member, Harold Hamm Diabetes Center, University of Oklahoma Health Sciences
Center
420 Engineering North, OSU, Stillwater, OK 74078
Ph: 405-744-5280
ashleefv@okstate.edu
Ford Versypt Lab Website <https://sites.google.com/site/ashleefordversypt/>

On Wed, May 2, 2018 at 8:59 PM, Kevin Mattheus Moerman <
notifications@github.com> wrote:

> @ashleefv <https://github.com/ashleefv> for the submitted works cited in
> your paper, are you able to post these on a pre-print server and add the
> URL's or DOI's to these references? e.g. https://engrxiv.org/ (you can
> check your rights for individual journals here
> <http://www.sherpa.ac.uk/romeo/index.php>)
>
