### 2018-05-11 #718 @sdelandtsheer
Hi, I have some minor remarks and a problem assessing the functionality:
- "scienfic" should be corrected in the paper and the .md
- if typical simulation times are 5-60 minutes and given bees' circadian rhythm is it correct to mention "days"?
- a number of scripts call a specific local path: _alternateParameterEstimation.m_, _Distr_RunNTumble.m_, _EstimateTransProb_Spatial.m_, _inSilicoExp.m_. Is it possible to modify them or remove them if they are not used?

I had issues running the GUI. File "beefiles/beefiles/Beeimage.jpg" does not exist, so the app does not start. Furthermore, I could not run inSilicoExp_Working.m, and traced it to motionParams being a 3x3 matrix instead of a structure.
Could you please either correct these and make sure the repo version is working as intended, or maybe provide more detailed instruction how to use the app for educational purposes and run simulations?
Thanks,
