### 2018-05-11 4018099
removed tangential scripts not needed for the program
