+++ Estimate_TransProb_Spatial.m
@@ -1,252 +1,0 @@
-clear; clc; close all
-%
-% Change the COLONY ID to select which colony data you want to process
-Colony_ID = 4;
-% Bees are considered to be bumping onto each other when their distance is
-% below the "BeeBodyThreshold"
-BeeBodyThreshold = 0.01;
-%
-DistToDetect_Nest_OnOff = 1*BeeBodyThreshold;
-% Loads the relevant data file
-% The data file should be inside "..\data" folder
-% This relative location should be maintained for proper functioning
-file2load = strcat('data\allDataCol',num2str(Colony_ID),'.mat');
-load(file2load);
-%
-%%
-% Counts the total number of time-frames and bees in the dataset
-% Number of time-frames can be different between pre- and post- exposure
-% datasets
-Total_Frames = size(distMatPre,3);
-Number_of_Bees = size(distMatPre,1);
-%
-% Computes the activity levels in the individual bees
-preNest_MDFD = calculateActivityMatrix(preNest);
-actvty_Storage_Pre = (preNest_MDFD(:,:,5))';
-%
-%
-% Computes whether a pair of bees are bumping into each other
-bumped_Storage = zeros(Number_of_Bees,Total_Frames);
-for time_idx = 1:Total_Frames
-    %
-    current_Distance = distMatPre(:,:,time_idx);
-    bumped_Storage(:,time_idx) = bump(BeeBodyThreshold,current_Distance);
-    %
-end
-%
-% Stores the bump-indicators inside a vector
-bumped_Vector = bumped_Storage(:,1:Total_Frames-1);
-bumped_Vector = bumped_Vector(:);
-% Stores the pre-bump activity level (active or inactive) inside a vector
-act_pr_Vector = actvty_Storage_Pre(:,1:Total_Frames-1);
-act_pr_Vector = act_pr_Vector(:);
-% Stores the post-bump activity level (active or inactive) inside a vector
-act_fu_Vector = actvty_Storage_Pre(:,2:Total_Frames);
-act_fu_Vector = act_fu_Vector(:);
-% From the vectors of pre- and post-bump activity levels computes a
-% variable to identify the transition (one of the following four types:
-% 0->0, 0->1, 1->0, 1->1
-Activity_Transition_Indicator_Pre = 2*act_pr_Vector + act_fu_Vector;
-%
-%
-% Copies the Activity Transition Indicators into two variables
-% One for those close to the nest structures
-% And another one for those away from the nest structures
-Activity_Transition_OnNest_Pre = Activity_Transition_Indicator_Pre;
-Activity_Transition_OffNest_Pre = Activity_Transition_Indicator_Pre;
-%
-bumped_Vector_OnNest = bumped_Vector;
-bumped_Vector_OffNest = bumped_Vector;
-%
-[WeDontCareHere,ON_or_OFF_Nest] = calculatePortionOfTimeOnNest(preNest_MDFD, broodPre, DistToDetect_Nest_OnOff);
-% The entries of the vector "ON_or_OFF_Nest" are either 0 or 1. 
-% 0 -->> Away from the Nest 
-% 1 -->> Close to the Nest
-ON_or_OFF_Nest = ON_or_OFF_Nest';
-ON_or_OFF_Nest = ON_or_OFF_Nest(:,1:Total_Frames-1);
[…]
