+++ inSilicoExp.m
@@ -1,259 +1,0 @@
-%Load initial data
-%base = '/n/home04/jcrall/pesticideModel';
-base = '/Users/james/Documents/pesticideProject';
-
-runLabel = 'Jan9_master';
-addpath(base);
-%base = '/Users/james/Documents/pesticideProject';
-cd(base);
-colonyNumber = 1; % 4 colonies numbered 1 - 4
-colony = load(['data/allDataCol' num2str(colonyNumber) '.mat']);
-estimatedData = load(['data/Essential_Info_Col_' num2str(colonyNumber) '.mat']);
-brood = colony.broodPre;
-treatmentList = colony.orTagTreat;
-taglist = colony.tags; %Read out colony's list of marked bees
-
-optimization = 0; %Optimiization(1) or experiment(0)?
-
-%%
-%treatments = {'full', 'sp', 'soc', 'control', 'fullNoSp', 'controlNoSp'}; %Simulation types -
-%full is fully parameterized model (including treatment-specific effects
-%
-%  "sp" is sppontaneous motion parameters driven  differently between control
-%and treatment
-%
-% "soc" - inverse of "sp", only socially-modulated switching parameters are
-% affected
-%
-% 'act' - full activity switching model, but no group-specific
-% environmental attraction
-%
-% "control" - all "treatment" group values are set equal to control values
-%
-% "fullNoSp" - including all treatment effects, but average parameters
-% between on and off nest (wiping out spatial effects)
-%
-% 'ap' - include group-specific environmental attraction parameters, but
-% equivalent mobility-switching parameters for all
-%
-%  "controlNoSp" -
-
-treatments = {'control', 'sp', 'soc', 'act', 'ap', 'full'}; %Simulation types  for current run
-
-%% Set simulation parameters
-ntrt = length(treatments); %How many treatments?
-nrep = 1000; %How many simulated replicates to run for each condition??
-groups = 1:3;
-ngroups = numel(groups);
-samplesPerRun = 7200;
-burnIn = 200; %How many timesteps to ignore at the beginning of the simulation?
-
-totSamps = ntrt*nrep*numel(groups); %Total samples
-
-
-%% define motion parameters
-params = readtable('parameterEstimates.csv');
-%rng(1); %Set random number generation seed for replicability
-if optimization  == 1 %For setting motion parameters, are these being drawn from the empirical values (specific to
-    %individual and colony), or by reshuffling the distribution of values by cohort? (non-empirical)?
-    empirical = 1;
-    pre = 1;
-    motionParamsCurrent = generateMotionParamsObject(params, colonyNumber, taglist, pre, empirical);
-    
-elseif optimization == 0
-    disp('running a simulation - will resample transition matrixes for each timestep');
-    empirical = 0;
-    pre = 0;
-    motionParamsCurrent = generateMotionParamsObject(params, colonyNumber, taglist, pre, empirical); %Generate randomly sampled, treatment-specific parameters
-    motionParamsControl = generateMotionParamsObject(params, colonyNumber, taglist, pre, 2); %Generate randomly sampled control data
-    
-end
-% motionParamsCurrent
-% subplot(2,1,1);
-% boxplot(motionParamsCurrent.AtoI_Unbumped_OffNest, treatmentList == 3)
-% subplot(2,1,2);
-% boxplot(motionParamsControl.AtoI_Unbumped_OffNest, treatmentList == 3)
-
-%%
-parpool('local', 20);
-
-%%
-
-
-for i = 1:ntrt
-    treatment = treatments{i};
-    
-    % Initialize separate output structures for each treatment groups
-    g1Dat = zeros(nrep, 8);
-    g2Dat = zeros(nrep, 8);
[…]

[…]
