### 2018-05-11 85761c3
updated modelParamsFull to call generateMotionParamsObject_general script using parameterEstimates.csv instead of hard-coded 3x3 matrix
