### 2018-05-11 9bc8eed
updated Beeimage and data files to run the app
