### 2018-05-11 #718 @ashleefv
Replying to the bullets in order

   - corrected in paper.md and README.md
   - In the experimental measurements, the bees are videotaped for 1 hour
   periods on different days. Additionally, when they are dosed with
   pesticides, the data is collected 24 hours after dosing for a one hour
   period. Thus, we deem it appropriate to mention days among in the short
   time scales.
   - The mentioned scripts were removed as they are not needed for the main
   functionality of the model. They were included in development stages.
   - Regarding inSilicoExp_Working.m, this was used for our first more
   specific paper (under review). We updated the modelParams object when we
   generalized for modelBeeScript.m. I have updated this section of
   inSilicoExp_Working.m and a related changed in simulationOutputSpatial.m.
   Both seem to be working as expected now.
   - Regarding the GUI, the image file was updated and referenced properly
   and the corresponding data files were included. The GUI is now running as
   intended.

Ashlee N. Ford Versypt, Ph.D.
Assistant Professor, School of Chemical Engineering, Oklahoma State
University
Member, Harold Hamm Diabetes Center, University of Oklahoma Health Sciences
Center
420 Engineering North, OSU, Stillwater, OK 74078
Ph: 405-744-5280
ashleefv@okstate.edu
Ford Versypt Lab Website <https://sites.google.com/site/ashleefordversypt/>
