### 2018-05-24 #718 @sdelandtsheer
Hi,
- thanks for the correction in the paper, I now get your intention behind 'days' and find it valid.
- while the app starts, the following error is generated when running:
_Index exceeds matrix dimensions._
_Error in rules (line 73)_
    _paramcase = varargin{2};_
_Error in simulation_attraction (line 18)_
    _nestSimulationData(timestep+1,:,:) = rules(dt,nestSimulationData(timestep,:,:),..._
so what happens is 'rules' is being called with 13 arguments while from the code I understand that if it has more than 12, then it should have 14 or 15. It also seems that in 'simulation_attraction.m', the variable dt is never defined.
- optimizerShellSpatial: does not run for me, however when 'tags' is changed to 'colony.tags' in lines 39, 115, 205 the script works. Could you confirm this is ok and make the respective corrections?
- inSilicoExp_Working.m: does not work either, except for the first treatment ('full'), and it seems that the variable MotionParamsCurrent is expected to have 3 dimensions, while in effect it has only two. This generates the following error:
_Error using inSilicoExp_Working (line 146)_
_You cannot subscript a table using linear indexing (one subscript) or multidimensional indexing (three or
more subscripts). Use a row subscript and a variable subscript._
The script 'generateMotionParamsObject_general.m' however always generate a table (2 dimensions).

In addition, there are scripts in the base folder and in the 'beefile' folder that are named identically. Sometimes they contain different code, which might make things go wrong depending the specific Matlab install and configuration. It would be much better to give them a unique name.

I hope this helps and I wait for your reply. I use Matlab 2017a. Let me know if I am getting anything wrong.

Best regards,
