### 2018-05-31 #718 @ashleefv
I'm traveling through the end of the week and should be able to address
them over the weekend.

Ashlee N. Ford Versypt, Ph.D.
Assistant Professor, School of Chemical Engineering, Oklahoma State
University
Member, Harold Hamm Diabetes Center, University of Oklahoma Health Sciences
Center
420 Engineering North, OSU, Stillwater, OK 74078
Ph: 405-744-5280
ashleefv@okstate.edu
Ford Versypt Lab Website <https://sites.google.com/site/ashleefordversypt/>

On Wed, May 30, 2018 at 9:19 PM, Kevin Mattheus Moerman <
notifications@github.com> wrote:

> @ashleefv <https://github.com/ashleefv> how are you getting on? Can you
> respond to @sdelandtsheer <https://github.com/sdelandtsheer> 's comments?
> Thanks. 🤖
>
