### 2018-06-10 84f75b6
updated beefile folder names to include _app to make app versions of files have unique names
