### 2018-06-10 d6a3564
updated beefile folder names to include _app to make app versions of files have unique names
