+++ BeeAttractionApp/beefiles/rules_app.m
@@ -1,4 +1,4 @@
-function output = rules(dt,currentState,broodPosition,...
+function output = rules_app(dt,currentState,broodPosition,...
     emptyFoodPosition,fullFoodPosition,...
     AtoI_Unbumped,ItoA_Unbumped,AtoI_Bumped,ItoA_Bumped,velocityPDF,exposure_state,tags,varargin)
 
@@ -80,31 +80,31 @@ cutoffRadius = zeros(numBees,1);
 
 % Fraction within +/- currentVelocity in which new velocity samples are 
 % always acceptable
-velocityPerturbationAlwaysAccept = assignCohortParameters(COHORTvelocityPerturbationAlwaysAccept,velocityPerturbationAlwaysAccept,tags);
+velocityPerturbationAlwaysAccept = assignCohortParameters_app(COHORTvelocityPerturbationAlwaysAccept,velocityPerturbationAlwaysAccept,tags);
 
 % Lambda weights for strength of attraction to nestmates (other bees),
 % brood, empty food pots, and full food pots
-lambdaBrood = assignCohortParameters(COHORTlambdaBrood,lambdaBrood,tags);
-lambdaFullFood = assignCohortParameters(COHORTlambdaFullFood,lambdaFullFood,tags);
-lambdaEmptyFood = assignCohortParameters(COHORTlambdaEmptyFood,lambdaEmptyFood,tags);
+lambdaBrood = assignCohortParameters_app(COHORTlambdaBrood,lambdaBrood,tags);
+lambdaFullFood = assignCohortParameters_app(COHORTlambdaFullFood,lambdaFullFood,tags);
+lambdaEmptyFood = assignCohortParameters_app(COHORTlambdaEmptyFood,lambdaEmptyFood,tags);
 lambdaBees(:) = 1-(lambdaBrood+lambdaFullFood+lambdaEmptyFood);
 
-environmentalStimuliWeight = assignCohortParameters(COHORTenvironmentalStimuliWeight,environmentalStimuliWeight,tags);
+environmentalStimuliWeight = assignCohortParameters_app(COHORTenvironmentalStimuliWeight,environmentalStimuliWeight,tags);
 
-perturbationAngle = assignCohortParameters(COHORTperturbationAngle,perturbationAngle,tags);
+perturbationAngle = assignCohortParameters_app(COHORTperturbationAngle,perturbationAngle,tags);
 
 % This fraction of samples outside of the velocityPerturbationAlwaysAccept 
 % window will be allowed to deviate to a proposedVelocity from the sampled 
 % velocityPDF. Those outside this fraction and also outside the 
 % velocityPerturbationAlwaysAccept window will remain at their currentVelocity
-velocityPerturbationMightAcceptProb = assignCohortParameters(COHORTvelocityPerturbationMightAcceptProb,velocityPerturbationMightAcceptProb,tags);
+velocityPerturbationMightAcceptProb = assignCohortParameters_app(COHORTvelocityPerturbationMightAcceptProb,velocityPerturbationMightAcceptProb,tags);
 
 % Spatial resolution that constitutes the distance at which bees are
 % considered to overlap or bump into each other
-BeeBodyThreshold = assignCohortParameters(COHORTBeeBodyThreshold,BeeBodyThreshold,tags);
+BeeBodyThreshold = assignCohortParameters_app(COHORTBeeBodyThreshold,BeeBodyThreshold,tags);
 
 % Cutoff radius for neighborhood sensing influence
-cutoffRadius = assignCohortParameters(COHORTcutoffRadius,cutoffRadius,tags);
[…]
