### 2018-06-10 #718 @ashleefv
- while the app starts, the following error is generated when running:*Index
   exceeds matrix dimensions.*
   - *Error in rules (line 73)*
   *paramcase = varargin{2};*
   *Error in simulation_attraction (line 18)*
   *nestSimulationData(timestep+1,:,:) =
   rules(dt,nestSimulationData(timestep,:,:),...*
   so what happens is 'rules' is being called with 13 arguments while from
   the code I understand that if it has more than 12, then it should have 14
   or 15. It also seems that in 'simulation_attraction.m', the variable dt is
   never defined.
   - optimizerShellSpatial: does not run for me, however when 'tags' is
   changed to 'colony.tags' in lines 39, 115, 205 the script works. Could you
   confirm this is ok and make the respective corrections?
   - inSilicoExp_Working.m: does not work either, except for the first
   treatment ('full'), and it seems that the variable MotionParamsCurrent is
   expected to have 3 dimensions, while in effect it has only two. This
   generates the following error:
   *Error using inSilicoExp_Working (line 146)*

*You cannot subscript a table using linear indexing (one subscript) or
   multidimensional indexing (three or more subscripts). Use a row subscript
   and a variable subscript.*
   The script 'generateMotionParamsObject_general.m' however always
   generate a table (2 dimensions).
   - I  updated beefile folder script names to include _app at the end to
   make app versions of the files have unique names.
   - I am using MATLAB 2017b, and I don't think that we have anything that
   explicitly required 2017b vs. a or slightly older versions.

Ashlee N. Ford Versypt, Ph.D.
Assistant Professor, School of Chemical Engineering, Oklahoma State
University
Member, Harold Hamm Diabetes Center, University of Oklahoma Health Sciences
Center
420 Engineering North, OSU, Stillwater, OK 74078
Ph: 405-744-5280
ashleefv@okstate.edu
Ford Versypt Lab Website <https://sites.google.com/site/ashleefordversypt/>

On Wed, May 30, 2018 at 8:26 PM, Ashlee Ford Versypt <
notifications@github.com> wrote:

> I'm traveling through the end of the week and should be able to address
> them over the weekend.
>
> Ashlee N. Ford Versypt, Ph.D.
> Assistant Professor, School of Chemical Engineering, Oklahoma State
> University
> Member, Harold Hamm Diabetes Center, University of Oklahoma Health Sciences
> Center
> 420 Engineering North, OSU, Stillwater, OK 74078
> Ph: 405-744-5280
> ashleefv@okstate.edu
> Ford Versypt Lab Website <https://sites.google.com/site/ashleefordversypt/
> >
>
>
>
> On Wed, May 30, 2018 at 9:19 PM, Kevin Mattheus Moerman <
> notifications@github.com> wrote:
>
> > @ashleefv <https://github.com/ashleefv> how are you getting on? Can you
> > respond to @sdelandtsheer <https://github.com/sdelandtsheer> 's
> comments?
> > Thanks. 🤖
> >
> > —
> > You are receiving this because you were mentioned.
> > Reply to this email directly, view it on GitHub
> > <https://github.com/openjournals/joss-reviews/issues/718#issuecomment-
> 393371174>,
> > or mute the thread
> > <https://github.com/notifications/unsubscribe-auth/AMD-
> NuCFkjvEMvaJcckntgUseb-SM44Uks5t30UcgaJpZM4TwZcd>
> > .
> >
>
