### 2018-06-10 #718 @ashleefv
My email was sent prematurely by accident..

I  updated beefile folder script names to include _app at the end to make app versions of the files have unique names.

I am using MATLAB 2017b, and I don't think that we have anything that explicitly required 2017b vs. a or slightly older versions.

Other edits still in progress. Will submit again later today.
