### 2018-06-10 b554bb8
removed reference to nonexistent colony.tags
