### 2018-06-10 f55c595
removed inSilicoExp_Working as it was not general purpose
