+++ inSilicoExp_Working.m
@@ -1,209 +1,0 @@
-%Load initial data
-colonyNumber = 1; % 4 colonies numbered 1 - 4
-colony = load(['data/allDataCol' num2str(colonyNumber) '.mat']);
-estimatedData = load(['data/Essential_Info_Col_' num2str(colonyNumber) '.mat']);
-brood = colony.broodPre;
-treatmentList = colony.orTagTreat;
-%%
-treatments = {'full', 'sp', 'soc', 'control', 'fullNoSp', 'controlNoSp'}; %Simulation types -
-%full is fully parameterized model (including treatment-specific effects
-%
-%  "sp" is sppontaneous motion parameters driven  differently between control
-%and treatment
-%
-% "soc" - inverse of "sp", only socially-modulated switching parameters are
-% affected
-%
-% "control" - all "treatment" group values are set equal to control values
-%
-% "fullNoSp" - including all treatment effects, but average parameters
-% between on and off nest (wiping out spatial effects)
-%
-%  "controlNoSp" -
-
-%% Set simulation parameters
-ntrt = length(treatments); %How many treatments?
-nrep = 3; %How many simulated replicates to run for each condition??
-groups = 1:3;
-samplesPerRun = 1000;
-burnIn = 100; %How many timesteps to ignore at the beginning of the simulation?
-
-totSamps = ntrt*nrep*numel(groups); %Total samples
-
-
-%% Initialize output data structure
-outData = array2table(nan(totSamps, 7));
-outData.Properties.VariableNames = {'activity', 'distanceToNestmates', 'distanceToBrood', 'distanceToCenter', 'porTimeOnNest', 'simType', 'treatmentGroup'};
-
-%% define motion parameters
-
-%general motionParams used for modelBeeScript that replaces the
-%"optimization" script below
-params = readtable('parameterEstimates.csv');
-motionParamsFull = generateMotionParamsObject_general(params, treatmentList);
-
-optimization = 1; %are we running optimization trials, or experimental trials? Optimization trials will use
-% %empirically derived parameters combined for all groups pre-exposure,
-% %otherwise group-specific post-exposure parameters are implemented
- 
-% if optimization  == 1
-%     
-%     
-%     %On nest parameters
-%     AIU1 = .01272; %Active to inactive unbumped, group 1
-%     IAU1 = .0325; %Inactive to active, unbumped, group 1...
-%     AIB1 = .01003; %Active to inactive, bumped, group 1...
-%     IAB1 = .3728;
-%     
-%     motionParamsOnNest = [AIU1 AIU1 AIU1 AIU1; IAU1 IAU1 IAU1 IAU1; AIB1 AIB1 AIB1 AIB1; IAB1 IAB1 IAB1 IAB1];
-%     
-%     %Off nest parameters
-%     AIU1 = .02959; %Active to inactive unbumped
-%     IAU1 = .2022; %Inactive to active, unbumped, group 1...
-%     AIB1 = .01749; %Active to inactive, bumped, group 1...
-%     IAB1 = .35560;
-%     
-%     %For "full" model
-%     motionParamsOffNest = [AIU1 AIU1 AIU1 AIU1; IAU1 IAU1 IAU1 IAU1; AIB1 AIB1 AIB1 AIB1; IAB1 IAB1 IAB1 IAB1];
-%     
-%     %Combine on and off-nest parameters sets
-% %     motionParamsFull = cat(3,motionParamsOnNest, motionParamsOffNest);
-%     
-% elseif optimization == 0
-%     
-%     
-%     %On nest parameters
-%     AIU1 = .015; %Active to inactive unbumped, group 1
-%     AIU2 = .015; % group 2...
-%     AIU3 = .015; %Etc
-%     AIU4 = .027;
-%     
[…]
