### 2018-06-10 f9480fb
Update README.md
removed reference to inSilicoExp_Working as it was not general purpose
