### 2018-06-10 #718 @ashleefv
When the app starts on my computer I do not get an error. Perhaps it was due to the two instances of the filename rules.m (and other scripts) in the repository. Do you get any errors now that the app files all have unique names? Also, dt is defined in bee_setup_script_app.m.

Note that 'tags' and 'taglist' are two separate things. tags = colony.orTagTreat is a vector of 0, 1, 2, or 3 for each bee indicating the level of dosing treatment, while taglist = colony.tags is a vector of 4 digit indices for the bees, which is not actually used later (so I have deleted it).  When I ran it, I got an error message that colony.tags wasn't even defined so I deleted taglist = colony.tags from line 39. This was a legacy statement from when we had .mat files with more information in them. Some time ago we reduced the size to simplify to information that we need for the simulations. We had run our tests and production runs on optimizerShellSpatial with the richer .mat files, so the useless line 39 hadn't thrown errors for us in another directory. Now I've tested it with 2016b and 2017b where optimizerShellSpatial.m is in the beeNestABm repository and reads the proper .mat files in the data subdirectory.

Good catch on the inSilicoExp_Working.m. My colleague ran the other non-full scenario "treatment groups". I had just seen that it ran on the "full" treatment scenario without issue. I have decided that I will remove inSilicoExp_Working.m because the in silico experiment it runs (with an older version of hard-coded motion params) is very specific to one empirical study that we have under review. Our more general simulation tool that we expect for others to adopt or use directly is modelBeeScript.m. It was developed most recently with the goal of enabling reuse and extension. So the inSilicoExp_Working.m file has been deleted from the repository, and the README has been updated accordingly with the modelBeeScript listed first among the main files now as it should be the most relevant for users.
