### 2018-06-22 #718 @Kevin-Mattheus-Moerman
Great. It looks like we are nearly there. @ashleefv would you be able to incorporate/mention the figure in the text of the paper?
