### 2018-06-26 8722348
Update paper.md
