### 2018-06-26 a2a0c44
Update paper.md
