### 2018-06-26 #718 @ashleefv
The figure caption was typed in the paper.md file but doesn't show up when compiling. This has been moved into the text as two stand-alone sentences at the end of the paragraph to mention the figure in the text.
