### 2018-06-26 1d9cc93
Update paper.md
