### 2018-06-26 #718 @ashleefv
My last name is Ford Versypt, two words without a hyphen. Is there a way to make the auto citation at the bottom of the article to include the Ford part without including a hyphen?
