### 2018-06-26 #718 @arfon
> My last name is Ford Versypt, two words without a hyphen. Is there a way to make the auto citation at the bottom of the article to include the Ford part without including a hyphen?

There is, but I'll have to do this manually before finishing up the paper acceptance. Please remind me to do this when we're done with the review.
