### 2018-06-26 #718 @ashleefv
Ok. I'm done with all the requested edits at this point.
