### 2018-06-27 #718 @Kevin-Mattheus-Moerman
Thanks @ashleefv can you create a DOI of the archived reviewed software and post it here? You can use a service like [zenodo](https://zenodo.org/), see also https://guides.github.com/activities/citable-code/.
