### 2018-07-02 cb69e91
updating v 2.0 DOI
