### 2018-07-03 #718 @Kevin-Mattheus-Moerman
Thanks @ashleefv. @arfon we are good to accept here. Above is the DOI and the version is now at v2.0.
