### 2018-07-03 #718 @ashleefv
Can we still do this?

My last name is Ford Versypt, two words without a hyphen. Is there a way to
make the auto citation at the bottom of the article to include the Ford
part without including a hyphen?

There is, but I'll have to do this manually before finishing up the paper
acceptance. Please remind me to do this when we're done with the review.

Ashlee N. Ford Versypt, Ph.D.
Assistant Professor, School of Chemical Engineering, Oklahoma State
University
Member, Harold Hamm Diabetes Center, University of Oklahoma Health Sciences
Center
420 Engineering North, OSU, Stillwater, OK 74078
Ph: 405-744-5280
ashleefv@okstate.edu
Ford Versypt Lab Website <https://sites.google.com/site/ashleefordversypt/>
