### 2018-07-03 #718 @arfon
Sure thing. I've just updated the paper. This may take a few hours to show as updated on the JOSS site as the PDFs are cached.
